Added coin command to view coin statistics
